Cambios mayores, ya no se usa el texto original, solo las palabras extraídas.
