+++ Textos/verbos_irregulares_infinitivos.txt
@@ -38,8 +38,8 @@ leer
 lucir
 mover
 mullir
-oír
 oler
+oír
 pedir
 peinar
 placer
