+++ Textos/verbos_regulares_infinitivos.txt
@@ -11463,7 +11463,6 @@ weber
 wélter
 xerocopiar
 xerografiar
-yacer
 yaguar
 yantar
 yapar
