+++ herramientas.py
@@ -1,0 +1,4 @@
+import sys
+
+def porcentuar(path):
+    '''Calcula una estimación del porcentaje del lenguaje utilizado.'''
