+++ main.py
@@ -1,22 +1,18 @@
 import sys
-import palabras
 import conjugaciones
-import utilidades
+import herramientas
 
 def inicializacion():
-    # TODO: Hacer que esto se haga solo cuando es necesario
-    print('Extrayendo palabras...')
-    palabras.extraer(r'Textos\diccionario.txt')
-    print('Generando conjugaciones...')
+    # Se conjugan los verbos regulares
     conjugaciones.generar()
-    # TODO: Combinar palabras extraidas en un único archivo
 
 if __name__ == '__main__':
     inicializacion()
+    print('\n\n', end='')
     inp = ''
     while not (inp == '1'):
         print('1) Salir.')
         print('Seleccione una opcion: ', end='')
         inp = input()
-    match inp:
-        case '1': exit()
+    if inp == '1':
+        exit()
