--- utilidades.py
@@ -1,4 +1,0 @@
-import sys
-
-def porcentuar(path):
-    '''Calcula una estimación del porcentaje del lenguaje utilizado.'''
